+++ dvclive/data/base.py
@@ -1,0 +1,44 @@
+import abc
+
+from pathlib import Path
+from typing import Optional
+
+from dvclive.error import AlreadyLoggedError
+
+
+class Data(abc.ABC):
+
+    def __init__(self, name: str, output_folder: str) -> None:
+        self.name = name
+        self.output_folder: Path = Path(output_folder)
+        self._step: Optional[int] = None
+        self.val = None
+
+    @property
+    def step(self) -> int:
+        return self._step
+    
+    @step.setter
+    def step(self, val: int) -> None:
+        if val == self._step:
+            raise AlreadyLoggedError(self.name, val)
+        self._step = val
+
+    @property
+    @abc.abstractmethod
+    def output_path(self) -> Path:
+        pass
+    
+    @property
+    @abc.abstractmethod
+    def summary(self):
+        pass
+
+    @staticmethod
+    @abc.abstractmethod
+    def could_log(val: object) -> bool:
+        pass
+
+    @abc.abstractmethod
+    def dump(self, val, step):
+        pass
